@@ -1,6 +1,6 @@
 {
   "name": "wds",
-  "version": "0.20.0",
+  "version": "0.21.0",
   "author": "Harry Brundage",
   "license": "MIT",
   "bin": {
@@ -14,6 +14,7 @@
     "access": "public"
   },
   "types": "pkg/index.d.ts",
+  "type": "module",
   "main": "pkg/index.js",
   "files": [
     "pkg/*",
@@ -31,20 +32,22 @@
     "lint:eslint": "NODE_OPTIONS=\"--max-old-space-size=4096\" eslint --quiet --ext ts,tsx src",
     "lint:fix": "NODE_OPTIONS=\"--max-old-space-size=4096\" prettier --write --check \"src/**/*.{js,ts,tsx}\" && eslint --ext ts,tsx --fix src",
     "prerelease": "gitpkg publish",
-    "test": "jest"
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "engines": {
     "node": ">=16.0.0"
   },
   "dependencies": {
-    "@cspotcode/source-map-support": "^0.8.1",
     "@swc/core": "^1.7.28",
+    "@swc/helpers": "^0.5.13",
     "chokidar": "^3.5.3",
     "find-root": "^1.1.0",
     "find-yarn-workspace-root": "^2.0.0",
+    "fs-extra": "^11.2.0",
     "globby": "^11.1.0",
-    "graceful-fs": "^4.2.11",
     "lodash": "^4.17.20",
+    "oxc-resolver": "^1.12.0",
     "pkg-dir": "^5.0.0",
     "write-file-atomic": "^6.0.0",
     "yargs": "^16.2.0"
@@ -52,21 +55,18 @@
   "devDependencies": {
     "@gadgetinc/eslint-config": "^0.6.1",
     "@gadgetinc/prettier-config": "^0.4.0",
-    "@swc/jest": "^0.2.36",
     "@types/find-root": "^1.1.2",
-    "@types/graceful-fs": "^4.1.9",
-    "@types/jest": "^27.4.0",
+    "@types/fs-extra": "^11.0.4",
     "@types/lodash": "^4.14.194",
-    "@types/node": "^18.11.9",
+    "@types/node": "^22.7.4",
     "@types/write-file-atomic": "^4.0.3",
     "@types/yargs": "^15.0.14",
-    "eslint": "^8.40.0",
-    "eslint-plugin-jest": "^27.2.1",
-    "gitpkg": "^1.0.0-beta.2",
-    "jest": "^27.4.7",
+    "eslint": "^8.57.1",
+    "gitpkg": "github:airhorns/gitpkg#gitpkg-v1.0.0-beta.4-gitpkg-82083c3",
     "prettier": "^2.8.8",
-    "typescript": "^5.1.3",
+    "typescript": "^5.6.2",
+    "vitest": "^2.1.1",
     "zx": "^7.2.3"
   },
-  "packageManager": "pnpm@8.12.1+sha256.28ca61ece5a496148b73fabc9afb820f9c3fec4f55f04ce45a2cea0a5219f2e1"
+  "packageManager": "pnpm@8.12.0+sha256.553e4eb0e2a2c9abcb419b3262bdc7aee8ae3c42e2301a1807d44575786160c9"
 }
