@@ -1,6 +1,6 @@
-const { execSync } = require('child_process')
+import { execSync } from "child_process";
 
-module.exports = () => ({
+export default () => ({
   getTagName: (pkg) =>
     `${pkg.name}-v${pkg.version}-gitpkg-${execSync(
       'git rev-parse --short HEAD',
