@@ -1,4 +1,4 @@
 const leak = [];
 while (true) {
-  leak.push("consuming memory");
+  leak.push("consuming memory".repeat(10000));
 }
