@@ -1,8 +1,9 @@
-import { promises as fs } from "graceful-fs";
-import { defaults } from "lodash";
+import fs from "fs-extra";
+import _ from "lodash";
 import path from "path";
-import type { ProjectConfig } from "./Options";
 import { threadId } from "worker_threads";
+// @ts-expect-error see https://github.com/microsoft/TypeScript/issues/52529, can't import types from .cts to .ts files that are ESM
+import type { ProjectConfig } from "./Options";
 
 const logPrefix = `[wds pid=${process.pid} thread=${threadId}]`;
 export const log = {
@@ -12,13 +13,13 @@ export const log = {
   error: (...args: any[]) => console.error(logPrefix, ...args),
 };
 
-export const time = async <T>(run: () => Promise<T>) => {
+export async function time<T>(run: () => Promise<T>) {
   const time = process.hrtime();
   await run();
   const diff = process.hrtime(time);
 
   return (diff[0] + diff[1] / 1e9).toFixed(5);
-};
+}
 
 export const projectConfig = async (root: string): Promise<ProjectConfig> => {
   const location = path.join(root, "wds.js");
@@ -26,12 +27,14 @@ export const projectConfig = async (root: string): Promise<ProjectConfig> => {
   try {
     await fs.access(location);
   } catch (error: any) {
-    log.debug(`Not loading project config from ${location}, error encountered: ${error.message}`);
+    log.debug(`Not loading project config from ${location}`);
     return value;
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  const required = require(location);
+  let required = await import(location);
+  if (required.default) {
+    required = required.default;
+  }
   log.debug(`Loaded project config from ${location}`);
-  return defaults(required, value);
+  return _.defaults(required, value);
 };
